Move services in dedicated folder
