+++ inc/services/acf.php
@@ -1,6 +1,9 @@
 <?php
 
-namespace BEA\Theme\Framework;
+namespace BEA\Theme\Framework\Services;
+
+use BEA\Theme\Framework\Service;
+use BEA\Theme\Framework\Service_Container;
 
 /**
  * Class Acf
@@ -31,7 +34,8 @@ class Acf implements Service {
 	/**
 	 * @param Service_Container $container
 	 */
-	public function register( Service_Container $container ) {}
+	public function register( Service_Container $container ) {
+	}
 
 	/**
 	 * @return string
