+++ inc/services/favicons.php
@@ -1,6 +1,9 @@
 <?php
 
-namespace BEA\Theme\Framework;
+namespace BEA\Theme\Framework\Services;
+
+use BEA\Theme\Framework\Service;
+use BEA\Theme\Framework\Service_Container;
 
 /**
  * Class Favicons
