+++ inc/services/theme.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace BEA\Theme\Framework;
+namespace BEA\Theme\Framework\Services;
+use BEA\Theme\Framework\Service;
+use BEA\Theme\Framework\Service_Container;
 
 
 class Theme implements Service {
